Added Start/stop button + tested timer
Added a Start and stop button to command the reading of the data. Also
had to change some stuff around the reading thread to make it work
correctly with the start and stop.

Added an "easy" time.sleep inside the read thread in order to schedule
the thread. Not dynamic for the moment, it will be for the next commit